[Rename] BookInfoBox 컴포넌트 이름 오기 수정
